fleet-server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
pyyaml==6.0.1
websockets==12.0
python-multipart==0.0.9
docker==7.1.0
